requirements.txt
python-telegram-bot[ext]==21.*
aiohttp
aiodns
python-dotenv
